rename(): test 내 클래스 main 으로 이동 후 패키지 별로 구분
