+++ src/main/java/christmas/domain/order/AcceptedOrders.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.order;
 
 import java.util.ArrayList;
 import java.util.Collections;
