+++ src/main/java/christmas/domain/order/Order.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.order;
 
 import static christmas.domain.menu.MenuProduct.findMenuProductByName;
